+++ mapflow/dialogs/main_dialog.py
@@ -1,20 +1,19 @@
 import sys
-
 from pathlib import Path
 from typing import Iterable, Optional, List
 
 from PyQt5 import uic
 from PyQt5.QtCore import Qt, pyqtSignal
-from PyQt5.QtGui import QPalette, QStandardItemModel
-from PyQt5.QtWidgets import QWidget, QPushButton, QComboBox, QCheckBox, QTableWidgetItem
+from PyQt5.QtGui import QPalette
+from PyQt5.QtWidgets import QWidget, QPushButton, QCheckBox, QTableWidgetItem
 from qgis.core import QgsMapLayerProxyModel, QgsMapLayer, QgsSettings
 
+from . import icons
+from ..config import config
 from ..entity.billing import BillingType
 from ..entity.provider import ProviderInterface
-from ..schema.project import MapflowProject
 from ..functional import helpers
-from ..config import config
-from . import icons
+from ..schema.project import MapflowProject
 
 ui_path = Path(__file__).parent/'static'/'ui'
 
@@ -26,6 +25,7 @@ class MainDialog(*uic.loadUiType(ui_path/'main_dialog.ui')):
     # SIGNALS
     modelOptionsChanged = pyqtSignal()
     rasterSourceChanged = pyqtSignal()
+    metadataTableFilled = pyqtSignal()
 
     def __init__(self, parent: QWidget, settings: QgsSettings) -> None:
         """Plugin's main dialog."""
@@ -272,7 +272,7 @@ class MainDialog(*uic.loadUiType(ui_path/'main_dialog.ui')):
 
         if fill:
             self.fill_metadata_table(fill)
-            
+
     def show_wd_price(self,
                       wd_price: float,
                       wd_description: str,
@@ -396,6 +396,7 @@ class MainDialog(*uic.loadUiType(ui_path/'main_dialog.ui')):
                 self.metadataTable.setItem(row, col, table_item)
         # Turn sorting on again
         self.metadataTable.setSortingEnabled(True)
+        self.metadataTableFilled.emit()
 
     def setup_project_combo(self, projects: List[MapflowProject], current_position: int):
         self.projectsCombo.clear()